pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "django-solana-payments"
version = "0.1.0"
description = "A customizable plug-and-play Solana Pay integration for Django"
authors = [
    { name = "Artemooon", email = "artiom.logachiov@gmail.com" }
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "Django>=4.2",
    "solana>=0.36.10",
    "solders>=0.18.0",
    "stamina>=23.1.0",
    "httpx>=0.24.0",
    "cryptography>=46.0.3,<47.0.0",
]

[project.optional-dependencies]
drf = ["djangorestframework >= 3.14"]
dev = [
    "pytest>=9.0.2,<10.0.0",
    "pytest-django>=4.11.1,<5.0.0",
    "pre-commit>=4.0.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-ra -q"
